refactor countrypanel
