+++ web/src/components/countryhistorycarbongraph.jsx
@@ -10,8 +10,7 @@ import AreaGraph from './graph/areagraph';
 import { getCO2IntensityByMode } from '../helpers/zonedata';
 
 const prepareGraphData = (historyData, co2ColorScale, electricityMixMode) => {
-  if (!historyData || !historyData[0] || !historyData[0].hasData) {
-    // TODO: TEST
+  if (!historyData || !historyData[0] || historyData.every((d) => !d.isValid)) {
     // Incomplete data
     return {};
   }
